% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnd_bias.R
\name{estimate_bias_parallel}
\alias{estimate_bias_parallel}
\title{estimate_bias_parallel}
\usage{
estimate_bias_parallel(
  iters,
  sampleN,
  rr_voi.given.cv.status = 1.5,
  rr_voi.given.cv.status.min = NA,
  rr_voi.given.cv.status.max = NA,
  coverage_voi = 0.7,
  coverage_cv = 0.55,
  ve_cv = 0.4,
  ve_voi = 0.9,
  risk_cv.dx.given.no.vx = 0.0215,
  risk_voi.dx.given.no.vx = 0.029,
  weight = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1)
)
}
\arguments{
\item{iters}{Number of iterations simulation will run}

\item{sampleN}{The total number in the population}

\item{rr_voi.given.cv.status}{Risk ratio of vaccine of interest given confounding vaccination status.
\deqn{\frac{Pr(vaccine of interest | vaccinated with confounding vaccine)}{Pr(vaccine of interest | not vaccinated with confounding vaccine)}}{Pr(vaccine of interest | vaccinated with confounding vaccine) / Pr(vaccine of interest | not vaccinated with confounding vaccine}}

\item{coverage_voi}{Vaccine of interest coverage. Pr(vaccine of interest).}

\item{coverage_cv}{Confounding vaccine coverage. Pr(confounding vaccine).}

\item{ve_cv}{Effectiveness of confounding vaccine.
1-(Pr(confounding vaccine disease | vaccinated with confounding vaccine) / Pr(confounding vaccine disease | not vaccinated with confounding vaccine))}

\item{ve_voi}{Effectiveness of vaccine of interest (VOI).
1-(Pr(vaccine of interest disease | vaccinated with vaccine of interest) / Pr(vaccine of interest disease | not vaccinated with vaccine of interest))}

\item{risk_cv.dx.given.no.vx}{Risk of confounding vaccine disease among the unvaccinated. Pr(confounding vaccine disease | not vaccinated with confounding vaccine).}

\item{risk_voi.dx.given.no.vx}{Risk of vaccine of interest disease among the unvaccinated. Pr(vaccine of interest disease | not vaccineated with vaccine of interest).}

\item{weight}{Pr(confounding vaccine disease | control). A vector of numeric may be specified.}
}
\description{
estimate_bias_parallel
}
