% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnd_bias.R
\name{one_iter}
\alias{one_iter}
\title{One iteration of the simulation.}
\usage{
one_iter(
  sampleN,
  rr_voi.given.cv.status = 8,
  coverage_voi = 0.7,
  coverage_cv = 0.55,
  ve_cv = 0.4,
  ve_voi = 0.9,
  risk_cv.dx.given.no.vx = 0.05,
  risk_voi.dx.given.no.vx = 0.05,
  weight = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1),
  fixed = FALSE
)
}
\arguments{
\item{sampleN}{The total number in the population}

\item{rr_voi.given.cv.status}{Risk ratio of vaccine of interest given confounding vaccination status.
\deqn{\frac{Pr(vaccine of interest | vaccinated with confounding vaccine)}{Pr(vaccine of interest | not vaccinated with confounding vaccine)}}{Pr(vaccine of interest | vaccinated with confounding vaccine) / Pr(vaccine of interest | not vaccinated with confounding vaccine}}

\item{coverage_voi}{Vaccine of interest coverage. Pr(vaccine of interest).}

\item{coverage_cv}{Confounding vaccine coverage. Pr(confounding vaccine).}

\item{ve_cv}{Effectiveness of confounding vaccine.
1-(Pr(confounding vaccine disease | vaccinated with confounding vaccine) / Pr(confounding vaccine disease | not vaccinated with confounding vaccine))}

\item{ve_voi}{Effectiveness of vaccine of interest (VOI).
1-(Pr(vaccine of interest disease | vaccinated with vaccine of interest) / Pr(vaccine of interest disease | not vaccinated with vaccine of interest))}

\item{risk_cv.dx.given.no.vx}{Risk of confounding vaccine disease among the unvaccinated. Pr(confounding vaccine disease | not vaccinated with confounding vaccine).}

\item{risk_voi.dx.given.no.vx}{Risk of vaccine of interest disease among the unvaccinated. Pr(vaccine of interest disease | not vaccineated with vaccine of interest).}

\item{weight}{Pr(confounding vaccine disease | control). A vector of numeric may be specified.}
}
\description{
One iteration of the simulation.
}
